Added Saving Rounds!
Scrappy and probably inefficient but it works! Added a new function and some new code when the round is ended to add to local storage to save. I have to change a lot and ill probably expand a ton more on this but its 1:30am and I need to sleep.